fix: ensure new backing images are based on target of symlinks
If the disk image pointed at in the topology file is a symlink to a
different file, copy the target of the symlink for the new backing image
and not the actual symlink.

Also take into account situations where the disk image pointed at in the
topology file is a snapshot (or chain of snapshots) on top of a base
image. If so, rebase the snapshot chain onto the new backing image prior
to rebasing the final minimega snapshot.