fix: only bail on tunnel errors from node tunnel is being created on
When tunnels are created in a multi-node minimega mesh, all the mesh
nodes that aren't running the VM the tunnel is being created for will
return an error saying the VM couldn't be found. These can be safely
ignored, so this commit updates the tunnel creation code to weed out
errors generated by the node the VM resides on.